pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ragicamp"
version = "0.4.0"
description = "A modular framework for experimenting with RAG approaches"
readme = "README.md"
requires-python = ">=3.10,<3.14"  # vLLM 0.15+ requires Python 3.10-3.13
authors = [
    {name = "RAGiCamp Team"}
]
dependencies = [
    "numpy>=1.21.0", # numpy 2.x required by vLLM 0.15+ (via opencv-python-headless)
    "pandas>=1.3.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.10.0",
    "faiss-cpu>=1.7.4", # FAISS for similarity search (GPU version not yet compatible with B200)
    "sentence-transformers>=2.2.0",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "scikit-learn>=1.2.0",
    "accelerate>=0.20.0",
    "bitsandbytes>=0.41.0", # For 8-bit quantization
    "bert-score>=0.3.13",
    "bleurt-pytorch>=0.0.1", # PyTorch BLEURT (no TensorFlow dependency)
    # Experiment tracking and optimization
    "mlflow>=2.8.0", # Experiment tracking and model registry
    "ragas>=0.1.0", # RAG-specific evaluation metrics
    "optuna>=3.0.0", # Hyperparameter optimization
    # Configuration management
    "ipykernel>=6.31.0",
    "boto3>=1.42.39",
    "awscli>=1.30.0", # For B2 backup download script
    "vllm>=0.11.0,<0.15.0", # PagedAttention (0.11+ has GTE-Qwen2, <0.15 avoids flashinfer JIT issues)
    # Visualization
    "matplotlib>=3.5.0",
    "seaborn>=0.13.2",
    "wandb>=0.15.0",
    # RAG features
    "bm25s>=0.2.0", # BM25 for hybrid retrieval (vectorized batch search)
]
# NOTE: For GPU FAISS (once Blackwell/B200 is supported), manually install:
#   uv pip uninstall faiss-cpu && uv pip install faiss-gpu-cu12>=1.7.4
# NOTE: flash-attn has complex build requirements - install manually if needed:
#   pip install flash-attn --no-build-isolation

[project.scripts]
ragicamp = "ragicamp.cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.4.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ragicamp"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["ragicamp"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.uv.sources]
# bleurt-pytorch is from PyPI, no custom source needed
